Added function to assign data pointer to ShapeShifter
This feature can be used in application where memory bandwidth is limited.
It allows sending a preserialized message to a topic without any copy (other than the one in the kernel).
The only catch is that the data passed must not be changed until the message has been sent or undefined behavior will result.